+++ package.json
@@ -4,6 +4,7 @@
     "private": true,
     "dependencies": {
         "@next/bundle-analyzer": "^13.1.6",
+        "@noble/hashes": "^1.3.0",
         "@scure/base": "^1.1.1",
         "@scure/btc-signer": "^0.5.1",
         "@stacks/common": "^6.0.0",
