+++ src/utils/crypto.js
@@ -5,12 +5,6 @@ import { TESTNET, ASSUMED_TX_BYTES, BITCOIN_PRICE_API_URL, DEFAULT_DERIV_PATH } 
 import BIP32Factory from "bip32";
 import { ethers } from "ethers";
 import { ECPairFactory } from "ecpair";
-import { sha256 } from '@noble/hashes/sha256';
-import { hexToBytes, utf8ToBytes } from '@stacks/common';
-import { serializeTaprootSignature } from "bitcoinjs-lib/src/psbt/bip371";
-import { encode } from 'varuint-bitcoin';
-import { base64 } from '@scure/base';
-import SessionStorage, { SessionsStorageKeys } from "@services/session-storage";
 
 bitcoin.initEccLib(ecc);
 const bip32 = BIP32Factory(ecc);
@@ -46,120 +40,8 @@ export const TAPROOT_MESSAGE = (domain) =>
     // will switch to nosft.xyz once sends are implemented
     `Sign this message to generate your Bitcoin Taproot key. This key will be used for your ${domain} transactions.`;
 
-
-
-const bip322MessageTag = 'BIP0322-signed-message';
-
-// See tagged hashes section of BIP-340
-// https://github.com/bitcoin/bips/blob/master/bip-0340.mediawiki#design
-const messageTagHash = Uint8Array.from([
-  ...sha256(utf8ToBytes(bip322MessageTag)),
-  ...sha256(utf8ToBytes(bip322MessageTag)),
-]);
-
-export function hashBip322Message(message) {
-    return sha256(
-      Uint8Array.from([...messageTagHash, ...(typeof message === 'string' || message instanceof String ? utf8ToBytes(message) : message)])
-    );
-}
-
 // Used to prove ownership of address and associated ordinals
 // https://github.com/LegReq/bip0322-signatures/blob/master/BIP0322_signing.ipynb
-export async function signBip322MessageSimple() {
-    const message = await prompt("Please enter BIP322 message to sign", "");
-    let publicKey = SessionStorage.get(SessionsStorageKeys.NOSTR_PUBLIC_KEY);
-    const metamaskDomain = SessionStorage.get(SessionsStorageKeys.DOMAIN);
-    const nostrScript = getAddressInfo(toXOnly(publicKey.toString()));
-    const scriptPubkey = nostrScript.output;
-    const pubkey = nostrScript.pubkey;
-    
-    // Generate a tagged hash of message to sign
-    const prevoutHash = hexToBytes(
-      '0000000000000000000000000000000000000000000000000000000000000000'
-    );
-    const prevoutIndex = 0xffffffff;
-    const sequence = 0;  
-    const hash = hashBip322Message(message);
-    
-    // Create the virtual to_spend transaction
-    const commands = [0, Buffer.from(hash)];
-    const scriptSig = bitcoin.script.compile(commands);
-    const virtualToSpend = new bitcoin.Transaction();
-    virtualToSpend.version = 0;
-    virtualToSpend.locktime = 0;  
-    virtualToSpend.addInput(Buffer.from(prevoutHash), prevoutIndex, sequence, scriptSig);
-    virtualToSpend.addOutput(Buffer.from(scriptPubkey), 0);
-  
-    // Create the virtual to_sign transaction
-    const virtualToSign = new bitcoin.Psbt();
-    virtualToSign.setLocktime(0);
-    virtualToSign.setVersion(0);
-    const prevTxHash = virtualToSpend.getHash(); // or id?
-    const prevOutIndex = 0;
-    const toSignScriptSig = bitcoin.script.compile([106]);
-  
-    virtualToSign.addInput({
-      hash: prevTxHash,
-      index: prevOutIndex,
-      sequence: 0,
-      witnessUtxo: { script: Buffer.from(scriptPubkey, "hex"), value: 0 },
-      tapInternalKey: toXOnly(pubkey),
-    });
-    virtualToSign.addOutput({ script: toSignScriptSig, value: 0 });
-
-    const sigHash = virtualToSign.__CACHE.__TX.hashForWitnessV1(
-        0,
-        [virtualToSign.data.inputs[0].witnessUtxo.script],
-        [virtualToSign.data.inputs[0].witnessUtxo.value],
-        bitcoin.Transaction.SIGHASH_DEFAULT
-    );
-
-    const sign = await signSigHash(sigHash);    
-    virtualToSign.updateInput(0, {
-        tapKeySig: serializeTaprootSignature(Buffer.from(sign, "hex")),
-    });
-    virtualToSign.finalizeAllInputs();
-  
-    const toSignTx = virtualToSign.extractTransaction();
-
-    function encodeVarString(b) {
-      return Buffer.concat([encode(b.byteLength), b]);
-    }
-  
-    const len = encode(toSignTx.ins[0].witness.length);
-    const result = Buffer.concat([len, ...toSignTx.ins[0].witness.map(w => encodeVarString(w))]);
-
-    const toReturn = { virtualToSpend, virtualToSign: toSignTx, signature: base64.encode(result) };
-
-    const sig = `Your BIP322 signature is: ${toReturn.signature}`;
-    alert(sig);
-    console.log(toReturn);
-
-    return toReturn
-  }
-
-  export async function signPsbt() {
-    const message = await prompt("Please enter PSBT to sign", "");
-    const virtualToSign = bitcoin.Psbt.fromBase64(message);
-    const sigHash = virtualToSign.__CACHE.__TX.hashForWitnessV1(
-        0,
-        [virtualToSign.data.inputs[0].witnessUtxo.script],
-        [virtualToSign.data.inputs[0].witnessUtxo.value],
-        bitcoin.Transaction.SIGHASH_SINGLE | bitcoin.Transaction.SIGHASH_ANYONECANPAY
-    );
-    const sign = await signSigHash(sigHash);        
-    virtualToSign.updateInput(0, {
-        tapKeySig: serializeTaprootSignature(Buffer.from(sign, "hex")),
-    });
-
-    virtualToSign.finalizeAllInputs();
-
-    const toSignTx = virtualToSign.toHex();
-    const sig = `Your PSBT is: ${toSignTx}`;
-    alert(sig);
-    console.log(toSignTx);
-    return toSignTx;
-  }
 
 export const connectWallet = async (metamask) => {
     const { ethereum } = window;
@@ -192,31 +74,6 @@ export const connectWallet = async (metamask) => {
     }
     return window.nostr.getPublicKey();
 };
-
-export const signSigHash = async (sigHash) => {
-    const metamaskDomain = SessionStorage.get(SessionsStorageKeys.DOMAIN);
-
-    if (metamaskDomain) {
-        const { ethereum } = window;
-        let ethAddress = ethereum.selectedAddress;
-        if (!ethAddress) {
-            await ethereum.request({ method: "eth_requestAccounts" });
-            ethAddress = ethereum.selectedAddress;
-        }
-        const provider = new ethers.providers.Web3Provider(window.ethereum);
-        const toSign = `0x${Buffer.from(TAPROOT_MESSAGE(metamaskDomain)).toString("hex")}`;
-        const signature = await provider.send("personal_sign", [toSign, ethAddress]);
-        const seed = ethers.utils.arrayify(ethers.utils.keccak256(ethers.utils.arrayify(signature)));
-        const root = bip32.fromSeed(Buffer.from(seed));
-        const taprootChild = root.derivePath(DEFAULT_DERIV_PATH);
-        const privateKey=taprootChild.privateKey;
-        const keyPair = ECPair.fromPrivateKey(privateKey);
-        const tweakedSigner = tweakSigner(keyPair);
-        return tweakedSigner.signSchnorr(sigHash);
-    } else {
-        return await window.nostr.signSchnorr(sigHash.toString("hex"));
-    }
-}
 
 export function satToBtc(sat) {
     return Number(sat) / 10 ** 8;
